[FEAT] 학습완료 + 학습완료 단원 개수 제공

+++ src/main/java/jpabasic/pinnolbe/dto/User/UserInfoDto.java
@@ -14,7 +14,7 @@ public class UserInfoDto {
 //    private String name;
 //    private String role;
     private String childName;
-    private Long coin;
+    private int coin;
 
 
 }
